feat: Add redis dependency and store messages in Redis
This commit adds the redis dependency and modifies the package.json and package-lock.json files to include the redis package with version 4.7.0. It also introduces changes to the send.ts file, where messages are now stored in Redis using the redis client. The messages are associated with a user's name and have an expiry time of 24 hours.